New class Evaluation
Create a new class : Evaluation, containing all the variables evaluation and the
locks in two separate dictionnaries. This class act like a single dictionnary. If the key is a string,
it is used with the variables evaluation dictionnary, and if not, if it used
with the locks dictionnary.

Correct two eval methods yielding None instead of not yielding