Tokens replaces active nodes + randInt expression
The nodes now contain tokens which can move along the graph. Tokens
contain parameters so that those parameters move along with the token.
A boolean expression of an arc can recognize a token.
A node can contain multiple tokens.
A token moveing through an arc activates the consequences of that arc.
The consequences now include
- delete the token moving through the arc
- editing that token
- add a new token on any arc

New boolean expression : randInt(Variable,Int) which unify Variable
with any integer between 0 and Int-1.