Implicit IS expressions
Consequences and Formula can now be written with implicite
IS arithmetic expression. For example:

before : pPROP(X) and T is X+1 and pPROP2(T)
now : pPROP(X) and pPROP2(X+1)

Any arithmetic expression must contain only idenified variables
or constants, it must produce a valid type result. If it does not
the consequence is ignored or the formula returns always False.