Change the way the label items moves
There is now a difference between label items for nodes and arcs.
The node label items remains the same.
The arc label items are now attached to the center of the arc with a vector u such that the angle from the arc vector and u does not change
when the arc is moved.