Changes with graphics
Graphics (Sprite, text and shapes) look now more like properties.
For example, a sprite is now created with 'add gsName(...)' instead if 'add s(name, ...)'.
Graphics consequences can now contains arguments and kwarguments.
Every property of a graphic element is a kwargument. For example, we can add a line
between coordinates (100, 100) and (200, 200) with 'add glExample(x1=100,y1=100,x2=100,y2=100)

In the future, it will be possible to use those graphics properties in the triggers (as it is
done for the events and the properties).